# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RealWorld Conduit API",
    "description": "Conduit API — node-express-realworld-example-app (TypeScript 5.2 · Express 5.2 · Prisma 4.16 · PostgreSQL). Auth: `Authorization: Token <jwt>` or `Authorization: Bearer <jwt>` accepted by express-jwt HS256 helper. Error shapes vary per route and are documented without normalisation (see individual responses).",
    "version": "1.0.0",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "TokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token issued on registration or login. Both `Authorization: Token <jwt>` (RealWorld convention) and `Authorization: Bearer <jwt>` (standard) are accepted via the express-jwt `getToken` helper (HS256). A missing or invalid token on a `required` route triggers an `UnauthorizedError` from express-jwt and returns HTTP 401."
      }
    },
    "schemas": {
...
      "Profile": {
        "type": "object",
        "description": "User profile. Always returned inside a `{ profile }` envelope.",
        "properties": {
          "username": { "type": "string" },
          "bio": { "type": "string", "nullable": true },
          "image": { "type": "string", "nullable": true },
          "following": {
            "type": "boolean",
            "description": "`true` when the authenticated user follows this profile; `false` when unauthenticated or not following."
          }
        },
        "required": ["username", "following"]
      },
...
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "profile": { "$ref": "#/components/schemas/Profile" }
        },
        "required": ["profile"]
      },
...
      "UnauthorizedErrorResponse": {
        "type": "object",
        "description": "HTTP 401: express-jwt `UnauthorizedError` — token missing, expired, or malformed. Produced by the global error handler in `src/app.ts`.",
        "properties": {
          "status": { "type": "string", "enum": ["error"] },
          "message": { "type": "string" }
        },
        "required": ["status", "message"],
        "example": {
          "status": "error",
          "message": "missing authorization credentials"
        }
      },
...
      "EmptyErrorResponse": {
        "type": "object",
        "description": "Empty object `{}`. Used as the body for some HTTP 404 responses and `DELETE /api/articles/{slug}/comments/{id}` 200 success.",
        "additionalProperties": false,
        "properties": {}
      }
    }
  },
  "paths": {
...
    "/api/profiles/{username}": {
      "get": {
        "tags": ["Profiles"],
        "summary": "Get profile",
        "description": "Returns a user's public profile. Auth optional — `following` reflects the current user's follow state when authenticated.",
        "operationId": "getProfile",
        "security": [{ "TokenAuth": [] }, {}],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid token (only triggered when a token is provided and is malformed/expired).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorResponse" }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmptyErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/profiles/{username}/follow": {
      "post": {
        "tags": ["Profiles"],
        "summary": "Follow user",
        "description": "Follows a user. Auth required.",
        "operationId": "followUser",
        "security": [{ "TokenAuth": [] }],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile with `following: true`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Profiles"],
        "summary": "Unfollow user",
        "description": "Unfollows a user. Auth required.",
        "operationId": "unfollowUser",
        "security": [{ "TokenAuth": [] }],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile with `following: false`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
...
}